Now brokering with song IDs
SongCatalog can now return a song by its ID. This is used with the
changed backend to now return song IDs. We correlate this with the
master song list, since song IDs are now the requested metadata. Using
this we now display richer data in the request view. Later on we can use
this to make a rich searchable experience on the front end.